fixed total count on pagination

+++ src/main/java/mywild/wildguide/framework/web/Paged.java
@@ -9,7 +9,7 @@ public record Paged<T>(
     @NotNull
     int pageSize,
     @NotNull
-    long totalRecords,
+    int totalRecords,
     @NotNull
     List<T> data
 ) { }
